remove radarChart and old timelineView
