+++ iOS/UI/RootContainerViewController.swift
@@ -87,11 +87,11 @@ class RootContainerViewController: UIViewController {
     }
 
     func rerenderTopView() {
-//        let sp0: MainSwipablePage! = mainSwipablePage
-//        if !sp0.pages.isEmpty,
-//            let listPage = (sp0.pages[1] as? ShadowingListPage) {
-//            listPage.sentencesTableView.reloadData()
-//        }
+        let sp0: MainSwipablePage! = mainSwipablePage
+        if !sp0.pages.isEmpty,
+            let listPage = (sp0.pages[1] as? ShadowingListPage) {
+            listPage.topChartView.viewWillAppear()
+        }
         let sp1: InfiniteChallengeSwipablePage! = infiniteChallengeSwipablePage
         if !sp1.pages.isEmpty,
             let listPage = (sp1.pages[1] as? InfiniteChallengeListPage) {
